check_softwareupdate.sh: the state file is a parameter now for the cron variant.
The "dnf" does not seem to work for SELinux systems. Even with the appropriate
sudo rules present and magic restorecon commands run, we'd get tons of
SELinux alerts during a simple "dnf check-update". A comment in 1422381 stated:

 > The problems you are listing are because you need to have those scripts
 > relabeled to be in the nrpe domain. What you are wanting to do with that
 > policy would seem to allow nrpe to run all kinds of programs on your system
 > which you may not want. [...] In both cases these aren't general solutions
 > but ones you need to set up for your domain.
	-- Failed to create session: Bad message for NRPE check
	   https://bugzilla.redhat.com/show_bug.cgi?id=1422381

While we may want to revisit the "dnf" at a later date, we'll decided to
at least make the "cron" version work. But as the NRPE user doesn't have
permission to write to STATE=/var/tmp/nagios_updates.txt we even have to
make that file a parameter. That of course needs to be reflected in the
helper script. But even for the "cron" version some preparation is needed:

* (root) helper cron job, e.g.:
  0 */12 * * * /usr/local/sbin/check_softwareupdate_helper dnf /var/run/nrpe/nagios_updates.txt

* SELinux policy, for example:

  module local_softwareupdate 1.0;
  require {
        type var_run_t;
        type tmp_t;
        type nrpe_t;
        class file { create getattr open read unlink write };
        class dir { add_name remove_name write };
  }
  #============= nrpe_t ==============
  allow nrpe_t tmp_t:dir { add_name remove_name write };
  allow nrpe_t tmp_t:file { create unlink write };
  allow nrpe_t var_run_t:file { getattr open read };

With all that, check_softwareupdate.sh should work again on SELinux enabled
systems with the DNF package manager.

Signed-off-by: Christian Kujau <lists@nerdbynature.de>